Send heartbeat events when new messages are seen

+++ src/com/irccloud/android/EventsDataSource.java
@@ -12,7 +12,7 @@ import android.database.sqlite.SQLiteDatabase;
 public class EventsDataSource {
 	public class Event {
 		long eid;
-		int bid;
+		long bid;
 		int cid;
 		String type;
 		int highlight;
